Lots of good stuff!
Added a menu for the main window, and processing to recieve menu item notifications.
Added notification handling for child windows.
Centralized the storage of class pointers in window handles.
Started on the implementation of GETMINMAXINFO, which will restrict the minimum screen size.
Fixed up return values for functions that never failed.
Added entrypoints to specify min and max height/width.