Fixed up quite a bit with the framework, and got a demo of the PanelContainer working!
Currently there are some rendering issues with the borders of the splitters, so I've enabled a high visibility rendering mode to figure out what's wrong. It will look wierd until I fix the underlying issue.

The demo is interactive - toy around with it and let me know if anything breaks.